Move test components into tests folder
