+++ crt_portal/cts_forms/tests/test_all_section_assignments.py
